backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
aiosqlite==0.19.0
databases==0.8.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-multipart==0.0.6
